[aie] fix round-trip format for packet_flow
I8Attr are stored in char types, which get printed not as integers by default.
Also start adding roundtrip tests

+++ lib/AIEDialect.cpp
@@ -346,7 +346,7 @@ static void print(OpAsmPrinter &p, xilinx::AIE::PacketFlowOp op) {
 
   Region &body = op.ports();
   p << xilinx::AIE::PacketFlowOp::getOperationName();
-  p << '(' << op.ID() << ')';
+  p << '(' << (int)op.ID() << ')';
 
   p.printRegion(body,
                 /*printEntryBlockArgs=*/false,
